Mask sensitive key material in memory.
Certain secrets such as kek, cathedral secret and shared secret
are only used during small windows of time.

There is no need to keep these in plaintext in memory, so encrypt
them when they are not used.

This makes it a lot harder for them to be revealed easily via
memory leaks or memory dumps of your machine. Note that it
does not make it impossible, just much much harder.

An adversary who has knowledge of where the KYRKA context lies
will very quickly be able to reveal them.